chore: :truck: rename & reorder edene-pro-lab files
