+++ libs/edene-pro-lab/src/components/simple/CountrySelector/CountrySelector.stories.tsx
@@ -3,7 +3,10 @@ import { CountrySelector as CountrySelectorComponent } from './CountrySelector';
 
 export default {
     component: CountrySelectorComponent,
-    title: 'Components',
+    title: 'Components/Simple',
+    parameters: {
+        layout: 'fullscreen',
+    },
 } as Meta;
 
 const Template: Story = () => <CountrySelectorComponent />;
