+++ libs/edene-pro-lab/src/components/simple/Note/Note.tsx
