+++ libs/edene-pro-lab/src/components/simple/Note/index.ts
