+++ libs/edene-pro-lab/src/pages/forms/AccountSettings/AccountSettings.stories.tsx
@@ -2,7 +2,7 @@ import { Story, Meta } from '@storybook/react';
 import AccountSettingsForm from './index';
 
 export default {
-    title: 'EdeneProLab/Forms',
+    title: 'Pages/Forms',
     component: AccountSettingsForm,
 } as Meta;
